changed encryption path and refactored to one-liner
